requirements.txt
anyio==3.7.1
asgiref==3.8.1
asyncpg==0.25.0
cffi==1.17.1
click==8.1.8
cryptography==37.0.2
dnspython==2.7.0
ecdsa==0.19.0
email-validator==1.2.1
exceptiongroup==1.2.2
fastapi==0.75.2
greenlet==3.1.1
gunicorn==20.1.0
h11==0.14.0
idna==3.10
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.19
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.3
starlette==0.17.1
typing_extensions==4.12.2
uvicorn==0.17.6
